pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-transformer"
version = "0.1.0"
description = "Leakage-safe autoregressive candle forecasting and walk-forward backtest harness."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "requests>=2.28",
  "scikit-learn>=1.4",
  "scipy>=1.11",
  "torch>=2.2",
  "pyarrow>=15",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[project.scripts]
stx-backtest = "stock_transformer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
